# docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VEP C2M Service API",
        "description": "VEP C2M Service API",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [
        {
            "url": "",
            "description": "clouddev"
        },
        {
            "url": "",
            "description": "sit"
        },
        {
            "url": "",
            "description": "uat"
        },
        {
            "url": "",
            "description": "prd"
        }
    ],
    "components": {
        "schemas": {
...
            "ZOOMLicenseOptionalDto": {
                "type": "object",
                "properties": {}
            },
            "ZOOMLicenseMustDto": {
                "type": "object",
                "properties": {}
            },
...
        }
    },
    "paths": {
...
        "/reuseLicense/createZOOMLicenseRecord/{meetingId}": {
            "post": {
                "operationId": "ReuseLicenseController_createZOOMLicenseRecord",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ZOOMLicenseOptionalDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        },
...
        "/reuseLicense/releaseZOOMLicense": {
            "post": {
                "operationId": "ReuseLicenseController_releaseZOOMLicense",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ZOOMLicenseMustDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        },
...
    }
}